LSE-CDM/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.7=py38ha9d4c09_0
brotli-python=1.0.9=py38h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2024.8.30=py38h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
ffmpeg=4.3=hf484d3e_0
freetype=2.13.3=h4a9f257_0
gmp=6.3.0=h6a678d5_0
gnutls=3.6.15=he1e5248_0
idna=3.7=py38h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jpeg=9e=h5eee18b_3
kmeans-pytorch=0.3=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
lz4-c=1.9.4=h6a678d5_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.16=h5eee18b_0
packaging=24.1=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
pillow=10.4.0=py38h5eee18b_0
pip=24.2=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
pooch=1.7.0=py38h06a4308_0
pthread-stubs=0.3=h0ce48e5_1
pysocks=1.7.1=py38h06a4308_0
python=3.8.20=he870216_0
python-dateutil=2.9.0post0=py38h06a4308_2
python-tzdata=2025.2=pyhd3eb1b0_0
pytorch=1.13.0=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2024.1=py38h06a4308_0
readline=8.2=h5eee18b_0
requests=2.32.3=py38h06a4308_0
scipy=1.10.1=py38hf6e8229_1
setuptools=75.1.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.45.3=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h993c535_1
torchaudio=0.13.0=py38_cu117
torchvision=0.14.0=py38_cu117
tqdm=4.66.5=py38h2f386ee_0
typing_extensions=4.11.0=py38h06a4308_0
urllib3=2.2.3=py38h06a4308_0
wheel=0.44.0=py38h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
